chore: move lib.rs to monovertex.rs
