+++ rust/numaflow-core/src/transformer.rs
@@ -1,8 +1,8 @@
 use crate::error::Result;
 use crate::message::Message;
 use crate::shared::utc_from_timestamp;
-use crate::sourcetransform_pb::source_transform_client::SourceTransformClient;
-use crate::sourcetransform_pb::SourceTransformRequest;
+use crate::monovertex::sourcetransform_pb::source_transform_client::SourceTransformClient;
+use crate::monovertex::sourcetransform_pb::SourceTransformRequest;
 use tonic::transport::Channel;
 
 const DROP: &str = "U+005C__DROP__";
@@ -58,7 +58,7 @@ mod tests {
     use std::error::Error;
 
     use crate::shared::create_rpc_channel;
-    use crate::sourcetransform_pb::source_transform_client::SourceTransformClient;
+    use crate::monovertex::sourcetransform_pb::source_transform_client::SourceTransformClient;
     use crate::transformer::SourceTransformer;
     use numaflow::sourcetransform;
     use tempfile::TempDir;
@@ -102,7 +102,7 @@ mod tests {
         let mut client = SourceTransformer::new(SourceTransformClient::new(
             create_rpc_channel(sock_file).await?,
         ))
-        .await?;
+            .await?;
 
         let message = crate::message::Message {
             keys: vec!["first".into()],
@@ -166,7 +166,7 @@ mod tests {
         let mut client = SourceTransformer::new(SourceTransformClient::new(
             create_rpc_channel(sock_file).await?,
         ))
-        .await?;
+            .await?;
 
         let message = crate::message::Message {
             keys: vec!["second".into()],
